+++ spec/job/adapters/torque_spec.rb
@@ -205,12 +205,6 @@ describe OodCore::Job::Adapters::Torque do
       it { expect(pbs).to have_received(:submit_string).with(content, queue: nil, headers: {Account_Name: "my_account"}, resources: {}, envvars: {}) }
     end
 
-    context "with :min_phys_memory" do
-      before { adapter.submit(build_script(min_phys_memory: 1234)) }
-
-      it { expect(pbs).to have_received(:submit_string).with(content, queue: nil, headers: {}, resources: {mem: "1234KB"}, envvars: {}) }
-    end
-
     context "with :wall_time" do
       before { adapter.submit(build_script(wall_time: 94534)) }